requirements.txt
faiss-cpu==1.8.0.post1
googletrans==3.1.0a0
nltk==3.9.1
numpy==1.26.4
open_clip_torch==2.26.1
streamlit==1.38.0
torch==2.4.1+cpu
torchaudio==2.4.1+cpu
torchvision==0.19.1+cpu
tqdm==4.66.5
